+++ Backend/logics/functionLogics.php
@@ -1,0 +1,5 @@
+<?php
+function bmiCalc($w, $h){
+    $bmi = $w / pow($h, 2);
+    return $bmi;
+}
